refactor: rename theme-provider ThemeProvider
